fix: foreign keys migraciones

+++ database/migrations/2025_09_06_182509_create_alumno_ciclo_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('alumno_ciclo', function (Blueprint $table) {
             $table->id();
             $table->smallInteger('semestreCursado');
-            $table->integer('idAlumno')->index('idAlumno');
-            $table->integer('idCicloEscolar')->index('idCicloEscolar');
+            $table->unsignedBigInteger('idAlumno')->index('idAlumno');
+            $table->unsignedBigInteger('idCicloEscolar')->index('idCicloEscolar');
         });
     }
