+++ database/migrations/2025_09_06_182509_create_alumno_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('nia', 8)->unique('nia');
             $table->integer('numeroLista')->unique('numeroLista');
             $table->enum('situacion', ['ACTIVO', 'BAJA_TEMPORAL', 'BAJA_DEFINITIVA', 'EGRESADO']);
-            $table->integer('idPersona')->index('idPersona');
+            $table->unsignedBigInteger('idPersona')->index('idPersona');
             $table->softDeletes();
             $table->timestamps();
