+++ database/migrations/2025_09_06_182509_create_grupo_semestre_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('grupo_semestre', function (Blueprint $table) {
             $table->id();
-            $table->integer('idGrupo')->index('idGrupo');
-            $table->integer('idSemestre')->index('idSemestre');
+            $table->unsignedBigInteger('idGrupo')->index('idGrupo');
+            $table->unsignedBigInteger('idSemestre')->index('idSemestre');
         });
     }
