+++ database/migrations/2025_09_06_182509_create_persona_table.php
@@ -23,8 +23,8 @@ return new class extends Migration
             $table->enum('sexo', ['F', 'M']);
             $table->date('fechaNacimiento');
             $table->string('nss',50);
-            $table->integer('idDireccion')->index('idDireccion');
-            $table->integer('idCuenta')->index('idCuenta');
+            $table->unsignedBigInteger('idDireccion')->index('idDireccion');
+            $table->unsignedBigInteger('idCuenta')->index('idCuenta');
             $table->softDeletes();
             $table->timestamps();
         });
